requirements.txt
prompt-toolkit==3.0.36
PyYAML==6.0.1
questionary==2.0.1
wcwidth==0.2.13
